feat: separate components, make order dialog wrapper more generic
