+++ libs/deal-ticket/src/components/deal-ticket-manager.tsx
@@ -1,23 +1,12 @@
 import type { ReactNode } from 'react';
 import { useState } from 'react';
-import type { VegaTxState } from '@vegaprotocol/wallet';
-import {
-  OrderDialogWrapper,
-  VegaTransactionDialog,
-  VegaTxStatus,
-} from '@vegaprotocol/wallet';
+import { VegaTransactionDialog, VegaTxStatus } from '@vegaprotocol/wallet';
 import { DealTicket } from './deal-ticket';
 import type { DealTicketQuery_market } from './__generated__/DealTicketQuery';
 import { useOrderSubmit } from '@vegaprotocol/orders';
-import { OrderStatus, OrderType, Side } from '@vegaprotocol/types';
-import type { OrderEvent_busEvents_event_Order } from '@vegaprotocol/orders';
-import { Icon, Intent } from '@vegaprotocol/ui-toolkit';
-import {
-  addDecimalsFormatNumber,
-  formatLabel,
-  t,
-} from '@vegaprotocol/react-helpers';
-import { useEnvironment } from '@vegaprotocol/environment';
+import { OrderStatus } from '@vegaprotocol/types';
+import { Intent } from '@vegaprotocol/ui-toolkit';
+import { OrderFeedback } from './order-feedback';
 
 export interface DealTicketManagerProps {
   market: DealTicketQuery_market;
@@ -64,85 +53,6 @@ export const DealTicketManager = ({
   );
 };
 
-interface OrderFeedbackProps {
-  transaction: VegaTxState;
-  order: OrderEvent_busEvents_event_Order | null;
-}
-
-const OrderFeedback = ({ transaction, order }: OrderFeedbackProps) => {
-  const { VEGA_EXPLORER_URL } = useEnvironment();
-  if (!order) return null;
-
-  const title = getDialogTitle(order.status);
-
-  // Order on network but was rejected
-  if (order.status === OrderStatus.Rejected) {
-    return (
-      <OrderDialogWrapper
-        title={title}
-        icon={<Icon name="warning-sign" size={20} />}
-      >
-        <p data-testid="error-reason">
-          {order.rejectionReason &&
-            t(`Reason: ${formatLabel(order.rejectionReason)}`)}
-        </p>
-      </OrderDialogWrapper>
-    );
-  }
-
-  return (
-    <OrderDialogWrapper title={title} icon={<Icon name="tick" size={20} />}>
-      <div className="grid grid-cols-1 md:grid-cols-2 gap-8">
-        {order.market && (
-          <div>
-            <p>{t(`Market`)}</p>
-            <p>{t(`${order.market.name}`)}</p>
-          </div>
-        )}
-        <div>
-          <p>{t(`Status`)}</p>
-          <p>{t(`${order.status}`)}</p>
-        </div>
-        {order.type === OrderType.Limit && order.market && (
-          <div>
-            <p>{t(`Price`)}</p>
-            <p>
-              {addDecimalsFormatNumber(order.price, order.market.decimalPlaces)}
-            </p>
-          </div>
-        )}
-        <div>
-          <p>{t(`Amount`)}</p>
-          <p
-            className={
-              order.side === Side.Buy ? 'text-vega-green' : 'text-vega-red'
-            }
-          >
-            {`${order.side === Side.Buy ? '+' : '-'} ${order.size}
-            `}
-          </p>
-        </div>
-      </div>
-      <div className="grid grid-cols-1 gap-8">
-        {transaction.txHash && (
-          <div>
-            <p>{t('Transaction')}</p>
-            <a
-              className="underline break-words"
-              data-testid="tx-block-explorer"
-              href={`${VEGA_EXPLORER_URL}/txs/0x${transaction.txHash}`}
-              target="_blank"
-              rel="noreferrer"
-            >
-              {transaction.txHash}
-            </a>
-          </div>
-        )}
-      </div>
-    </OrderDialogWrapper>
-  );
-};
-
 const getDialogIntent = (status?: OrderStatus) => {
   if (!status) {
     return Intent.None;
@@ -158,20 +68,5 @@ const getDialogIntent = (status?: OrderStatus) => {
       return Intent.Danger;
     default:
       return Intent.Success;
-  }
-};
-
-const getDialogTitle = (status: OrderStatus) => {
-  switch (status) {
-    case OrderStatus.Active:
-      return 'Order submitted';
-    case OrderStatus.Filled:
-      return 'Order filled';
-    case OrderStatus.PartiallyFilled:
-      return 'Order partially filled';
-    case OrderStatus.Parked:
-      return 'Order parked';
-    default:
-      return 'Submission failed';
   }
 };
