+++ libs/orders/src/lib/components/order-list/order-edit-dialog.tsx
@@ -10,7 +10,9 @@ import {
   InputError,
   Button,
   Dialog,
+  Icon,
 } from '@vegaprotocol/ui-toolkit';
+import { VegaTransactionDialogWrapper } from '@vegaprotocol/wallet';
 import { useForm } from 'react-hook-form';
 import type { OrderFields } from '../order-data-provider';
 
@@ -47,56 +49,64 @@ export const OrderEditDialog = ({
   if (!order) return null;
 
   return (
-    <Dialog open={isOpen} onChange={onChange} title={t('Edit order')}>
-      <div className="grid grid-cols-1 md:grid-cols-3 gap-8">
-        {order.market && (
+    <Dialog open={isOpen} onChange={onChange}>
+      <VegaTransactionDialogWrapper
+        title={t('Edit order')}
+        icon={<Icon name="edit" />}
+      >
+        <div className="grid grid-cols-1 md:grid-cols-3 gap-8">
+          {order.market && (
+            <div>
+              <p className={headerClassName}>{t(`Market`)}</p>
+              <p>{t(`${order.market.name}`)}</p>
+            </div>
+          )}
+          {order.type === OrderType.Limit && order.market && (
+            <div>
+              <p className={headerClassName}>{t(`Current price`)}</p>
+              <p>
+                {addDecimalsFormatNumber(
+                  order.price,
+                  order.market.decimalPlaces
+                )}
+              </p>
+            </div>
+          )}
           <div>
-            <p className={headerClassName}>{t(`Market`)}</p>
-            <p>{t(`${order.market.name}`)}</p>
-          </div>
-        )}
-        {order.type === OrderType.Limit && order.market && (
-          <div>
-            <p className={headerClassName}>{t(`Last price`)}</p>
-            <p>
-              {addDecimalsFormatNumber(order.price, order.market.decimalPlaces)}
+            <p className={headerClassName}>{t(`Amount remaining`)}</p>
+            <p
+              className={
+                order.side === 'Buy'
+                  ? 'text-dark-green dark:text-vega-green'
+                  : 'text-red dark:text-vega-red'
+              }
+            >
+              {order.side === 'Buy' ? '+' : '-'}
+              {order.size}
             </p>
           </div>
-        )}
-        <div>
-          <p className={headerClassName}>{t(`Amount remaining`)}</p>
-          <p
-            className={
-              order.side === 'Buy'
-                ? 'text-dark-green dark:text-vega-green'
-                : 'text-red dark:text-vega-red'
-            }
-          >
-            {order.side === 'Buy' ? '+' : '-'}
-            {order.size}
-          </p>
         </div>
-      </div>
 
-      <div className="grid grid-cols-1 md:grid-cols-2 gap-8 py-12">
-        <form onSubmit={handleSubmit(onSubmit)} data-testid="edit-order">
-          <FormGroup label={t('Entry price')} labelFor="entryPrice">
-            <Input
-              {...register('entryPrice', { required: t('Required') })}
-              id="entryPrice"
-              type="text"
-            />
-            {errors.entryPrice?.message && (
-              <InputError intent="danger" className="mt-4">
-                {errors.entryPrice.message}
-              </InputError>
-            )}
-          </FormGroup>
-          <Button variant="primary" type="submit">
-            {t('Update')}
-          </Button>
-        </form>
-      </div>
+        <div className="grid grid-cols-1 md:grid-cols-2 gap-8 py-12">
+          <form onSubmit={handleSubmit(onSubmit)} data-testid="edit-order">
+            <FormGroup label={t('Entry price')} labelFor="entryPrice">
+              <Input
+                {...register('entryPrice', { required: t('Required') })}
+                id="entryPrice"
+                type="text"
+              />
+              {errors.entryPrice?.message && (
+                <InputError intent="danger" className="mt-4">
+                  {errors.entryPrice.message}
+                </InputError>
+              )}
+            </FormGroup>
+            <Button variant="primary" type="submit">
+              {t('Update')}
+            </Button>
+          </form>
+        </div>
+      </VegaTransactionDialogWrapper>
     </Dialog>
   );
 };
