+++ libs/wallet/src/vega-transaction-dialog/index.ts
@@ -1,1 +1,2 @@
 export * from './vega-transaction-dialog';
+export * from './vega-transaction-dialog-wrapper';
